+++ src/Data/Morpheus/Types/Internal/Resolving.hs
@@ -30,9 +30,10 @@ module Data.Morpheus.Types.Internal.Resolving
     , FieldRes
     , WithOperation
     , PushEvents(..)
-    , getContext
     , runDataResolver
     , subscribe
+    , Context(..)
+    , unsafeInternalContext
     )
 where